GH-49 :: refactoring
