+++ src/TrainingGuides.Admin/Pages/ProjectSettings/WebChannelSettings/SeoSettingsEditPage.cs
@@ -1,6 +1,6 @@
 using Kentico.Xperience.Admin.Base;
 using TrainingGuides.ProjectSettings;
-using TrainingGuides.Admin.ProjectSettings;
+using TrainingGuides.Admin.ProjectSettings.WebChannelSettings;
 using Kentico.Xperience.Admin.Base.Forms;
 using CMS.DataEngine;
 using Microsoft.Extensions.Localization;
@@ -13,7 +13,7 @@ using Microsoft.Extensions.Localization;
     templateName: TemplateNames.EDIT,
     order: 0)]
 
-namespace TrainingGuides.Admin.ProjectSettings;
+namespace TrainingGuides.Admin.ProjectSettings.WebChannelSettings;
 
 public class SeoSettingsEditPage : InfoEditPage<SeoSettingsInfo>
 {
@@ -61,7 +61,7 @@ public class SeoSettingsEditPage : InfoEditPage<SeoSettingsInfo>
         return base.ConfigurePage();
     }
 
-    protected async override Task<ICommandResponse> GetSubmitSuccessResponse(SeoSettingsInfo savedInfoObject, ICollection<IFormItem> items)
+    protected override async Task<ICommandResponse> GetSubmitSuccessResponse(SeoSettingsInfo savedInfoObject, ICollection<IFormItem> items)
     {
         var result = new EditPageSuccessFormSubmissionResult()
         {
